requirements.txt
paho-mqtt~=2.1.0
Flask~=3.1.0